Fix compiler warning: implicit conversion loses integer precision
src/gsb_account_property.c:1168:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_1 = g_utf8_pointer_to_offset ( model, ptr_1 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1170:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_2 = g_utf8_pointer_to_offset ( model, ptr_2 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1183:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_1 = g_utf8_pointer_to_offset ( model, ptr_1 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1187:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_2 = g_utf8_pointer_to_offset ( model, ptr_2 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1204:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_1 = g_utf8_pointer_to_offset ( model, ptr_1 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1208:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_2 = g_utf8_pointer_to_offset ( model, ptr_2 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1222:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_1 = g_utf8_pointer_to_offset ( model, ptr_1 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1226:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_2 = g_utf8_pointer_to_offset ( model, ptr_2 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1238:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_1 = g_utf8_pointer_to_offset ( model, ptr_1 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/gsb_account_property.c:1242:22: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        pos_char_2 = g_utf8_pointer_to_offset ( model, ptr_2 );
                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/gsb_account_property.c
@@ -1116,8 +1116,8 @@ gboolean gsb_account_property_iban_set_bank_from_iban ( gchar *iban )
     gchar *ptr_1;
     gchar *ptr_2;
     gchar *code;
-    gint pos_char_1;
-    gint pos_char_2;
+    glong pos_char_1;
+    glong pos_char_2;
     gint bank_number;
     gint current_account;
     gunichar c;
